pyproject.toml
[project]
name = "cellstream"
version = "0.1.0"
description = "Microscopy and single-cell signal processing analysis tools"
authors = [{ name = "Your Organization" }]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "tifffile",
    "nd2",
    "progressbar2",
    "ssqueezepy",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
